refactor(name): Changes the name and structure
