+++ bower.json
@@ -1,4 +1,4 @@
 {
-  "name": "ember-computed-decorators",
+  "name": "ember-decorators",
   "dependencies": {}
 }
