+++ index.js
@@ -2,5 +2,5 @@
 'use strict';
 
 module.exports = {
-  name: 'ember-computed-decorators'
+  name: 'ember-decorators'
 };
