+++ tests/dummy/config/environment.js
@@ -44,7 +44,7 @@ module.exports = function(environment) {
 
   if (environment === 'production') {
     ENV.locationType = 'hash';
-    ENV.rootURL = '/ember-computed-decorators/';
+    ENV.rootURL = '/ember-decorators/';
 
   }
